Move logic in a different file
