+++ packages/bff/package.json
@@ -37,13 +37,14 @@
     "axios": "1.7.2",
     "dotenv-flow": "4.1.0",
     "express": "4.19.2",
-    "pagopa-interop-commons": "workspace:*",
+    "fast-xml-parser": "^4.4.0",
+    "mime": "4.0.4",
+    "pagopa-interop-agreement-lifecycle": "workspace:*",
     "pagopa-interop-api-clients": "workspace:*",
+    "pagopa-interop-commons": "workspace:*",
     "pagopa-interop-models": "workspace:*",
     "pagopa-interop-selfcare-v2-client": "workspace:*",
-    "pagopa-interop-agreement-lifecycle": "workspace:*",
     "qs" : "6.12.3",
-    "fast-xml-parser": "^4.4.0",
     "ts-pattern": "5.2.0",
     "uuid": "10.0.0",
     "yaml": "2.4.5",
